Fix aggregated values for None samples
If the samples in the protobuf of the aggregated response has the value
not set (e.g. when the service sends None values), the generated python
code falls back to zeroes when the field is accessed. This case needs to
be handled explicitly by the client, which now sets these values to NaN.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>